# openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "CalmKit API",
    "description": "Mental wellness app API with emotion logging, AI coaching, anchor phrases, and guided routines",
    "version": "1.0.0"
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": ["http"],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "x-auth-token",
      "in": "header",
      "description": "JWT token for authentication"
    }
  },
  "paths": {
...
    "/api/routines/": {
      "get": {
        "tags": ["Routines"],
        "summary": "Get all routines",
        "description": "",
        "parameters": [
          {
            "name": "x-auth-token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": ["Routines"],
        "summary": "Create a new routine",
        "description": "",
        "parameters": [
          {
            "name": "x-auth-token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Routine data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Deep Breathing"
                },
                "description": {
                  "type": "string",
                  "example": "A calming breathing exercise"
                },
                "icon": {
                  "type": "string",
                  "example": "🫁"
                },
                "tags": {
                  "type": "array",
                  "example": ["Breathing", "Quick"],
                  "items": {
                    "type": "string"
                  }
                },
                "steps": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string",
                        "example": "Prepare"
                      },
                      "text": {
                        "type": "string",
                        "example": "Find a comfortable position"
                      },
                      "duration": {
                        "type": "number",
                        "example": 30
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/user/me": {
      "get": {
        "tags": ["User"],
        "summary": "Get current user info",
        "description": "",
        "parameters": [
          {
            "name": "x-auth-token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/user/avatar": {
      "put": {
        "tags": ["User"],
        "summary": "Update avatar color",
        "description": "",
        "parameters": [
          {
            "name": "x-auth-token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Avatar data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "color": {
                  "type": "string",
                  "example": "#4A9093"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
...
  "security": [
    {
      "bearerAuth": []
    }
  ]
}